+++ src/exporters/index.ts
@@ -1,4 +1,5 @@
 export * from "./arrow.js";
+export * from "./drawio.js";
 export * from "./excalidraw.js";
 export * from "./svg.js";
 export * from "./types.js";